refactor: test 분리
