--- core/src/test/java/kitchenpos/domain/DomainTest.java
@@ -1,8 +1,0 @@
-package kitchenpos.domain;
-
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-public abstract class DomainTest {
-}
