+++ menu/src/test/java/kitchenpos/menu/domain/MenuProductsTest.java
@@ -1,16 +1,17 @@
-package kitchenpos.domain.menu;
+package kitchenpos.menu.domain;
 
-import kitchenpos.domain.DomainTest;
-import kitchenpos.menu.domain.MenuProducts;
-import org.assertj.core.api.SoftAssertions;
-import org.junit.jupiter.api.Test;
 import kitchenpos.product.domain.Product;
+import org.assertj.core.api.SoftAssertions;
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
+import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class MenuProductsTest extends DomainTest {
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
+class MenuProductsTest {
     @Test
     void create_menu_products() {
         // given
